Update Node.java
changed the way new weights are formed on a new node.  only half of weights by chance are non zero.  if all weights turn up zero it tries again.  Threshold value is half of the highest possible positive value.  if no positive value is possible the threshold is half of the negative value.