Order of Operations
I was trying to expand the inputs before they were in there.  I've corrected that and now Inputs get expanded and Networks get created at the beginning of Trial.run() instead of in the Trial constructor.  This is because using the super constructor mucked up when things are made in the sub class.  Still haven't gotten to the bottom of this though.  HISTORY IS STILL FUCKED UP!  arg