StateVar --> BigDecimal
yep.  I was going to make auto conversions on the setters, but then I realized as soon as I pass a double as a parameter, it's already fucked.  Then I started thinking that I really should make all the inputs and outputs be BigDecimals.  then I thought FUCK THAT, this is good enough for a little project I'm making for myself that I am sure will not suceed (BTC Trading indicator).  Will be good to gain exp with BigDecimal though.