Trial.load() debugged
it is debugged, specifically for ColorTrial.  However when a trial is restarted, it will start on the right cycle but with a much lower score... which defeats the whole purpose.  I guess I'll need to debug that later.  It's harder now, as I only grab an hour here or there once a week.  kids